requirements.txt
black==19.10b0
bottle==0.12.18
google-cloud-datastore==1.10.0
google-cloud-logging==1.14.0
google-cloud-pubsub==1.3.0
google-cloud-storage==1.26.0
pipdeptree==0.13.2
pylint==2.4.4
python-dotenv==0.11.0
seaborn==0.10.0
vulture==1.3
